Update make file

+++ CSources/Factorial.c
@@ -1,5 +1,5 @@
-int factorial(int n) {
-  int result = 1;
+long factorial(long n) {
+  long result = 1;
   for(int i = 1;i <= n;i++) {
     result = result * i;
   }
